fix: lazy load types
Move ModelSignatureDict to _internal.types to fix import
