+++ src/bentoml/types.py
@@ -1,4 +1,4 @@
 from ._internal.models.model import ModelSignature
-from ._internal.models.model import ModelSignatureDict
+from ._internal.types import ModelSignatureDict
 
 __all__ = ["ModelSignature", "ModelSignatureDict"]
